stripe/requirements.txt
# Function dependencies, for example:
# package>=version
pandas>=1.2.2
google-cloud-storage==1.44.0
stripe>=4.2.0
